[Jaeyoung] <feat> 찜 하기 / 찜 취소 메소드 가구현(Token 가져오는 부분을 제외한)

+++ src/main/java/com/hanghae/mungnayng/domain/item/dto/ItemResponseDto.java
@@ -24,8 +24,8 @@ public class ItemResponseDto {
     private String itemCategory;
     private String itemImgs;
     private String location;
-    private Long commentCnt;
-    private Long zzimCnt;
+    private int commentCnt;
+    private int zzimCnt;
     private int viewCnt;
     private Long purchasePrice;
     private Long sellingPrice;
